Name changes, global functions
Moved functions back into core and called them in the controller. Also
moved the page construction in to the init function so there aren’t any
functions in view based files
