+++ app/controllers/index.php
@@ -1,0 +1,6 @@
+<?php
+
+// get the posts
+$posts = get_all_posts();
+
+?>
