+++ app/partials/header.php
@@ -1,4 +1,3 @@
-<?php include('core/functions.php');?>
 <!doctype html>
 <html lang="en">
 <head>
